Move sweetalert logic to action creator
